last part of the project toDo list done
