+++ src/task.js
@@ -1,0 +1,5 @@
+import { getLocalStorage } from './modules/storage.js';
+
+const toDoTasks = getLocalStorage();
+
+export default toDoTasks;
